kube-spawn: enlarge storage pool for each cloned image
Enlarge the storage pool that is actually a btrfs volume backed by
/var/lib/machines.raw. Every btrfs volume would need more space for its
metadata as much as the volume grows through CoW operations.
/var/lib/machines should have more space than its original image size.
Otherwise /var/lib/machines could become full when files start to be
written to nspawn containers.

Unfortunately it's a temporary workaround. How much space we should add
more to the image might depend on estimation during run-time operations.
In the long run, systemd itself should be able to reserve more space
for the storage pool, every time when it pulls an image to store in
the pool.

Fixes https://github.com/kinvolk/kube-spawn/issues/66