kube-spawn: ensure that system requirements are satisfied
Ensure that the system requirements are satisfied for starting
kube-spawn. It's just like running the commands below:

  # modprobe overlay
  # modprobe nf_conntrack
  # echo "131072" > /sys/module/nf_conntrack/parameters/hashsize

It was originally done by scripts/vagrant-mod-env.sh. However it's
actually hard for end-users to take care of the requirements prior
to running kube-spawn. So allow kube-spawn to do the initialization
by itself a bit actively.

Fixes https://github.com/kinvolk/kube-spawn/issues/69