cnispawn, nspawntool: check GOPATH, CNI_PATH in a flexible way
Check for availability of $GOPATH and $CNI_PATH in a flexible way.
If a given path is not a valid path to a directory, fall back to
the next possible path candidate. If it still fails, print out
warning. That way, users could get chances to fix the paths.

Partly covers https://github.com/kinvolk/kube-spawn/issues/81