pkg/bootstrap: check for Container Linux version
As kube-spawn needs Container Linux 1478.0.0 or newer, we need to check
for the version.

If the version is too low, it fails and recommends users to remove the
old local image, so that the next run of `kube-spawn` will download a
new image automatically.

Fixes https://github.com/kinvolk/kube-spawn/issues/114