kube-spawn machine naming: use kubespawn%d without dashes
The name of the machine given to machinectl (e.g. kube-spawn-0) is also
used to set the cgroup path. Example:
/machine.slice/machine-kube\x2dspawn\x2d0.scope

Having dashes make things more complicated because they are escaped by
systemd to '\x2d' in the cgroup path.

When adding the support for cgroup-per-qos in kube-spawn, we will need
to pass the parameter --cgroup-root= to the Kubelet. With the '\x2d'
escape sequence, we have now a backslash. That would need to be escaped
when writing the kubelet service file.

So instead of over-complicating things with a double-escaping sequence,
let's just get rid of the dashes.