cmd/kube-spawn: fix a wrong ContainerRuntime for bootstrap.sh
For the ContainerRuntime parameter of bootstrap.sh, we need to pass an
explicit string, either "rktlet" or "docker". Without this fix,
bootstrap.sh will silently fail in case of rkt, because there's no
rkt.service to enable. As a result, the rest of commands in bootstrap.sh
will not run at all.