pkg/config: add groups option to token create for kubeadm v1.8
Since kubeadm v1.8 or newer introduced a new option --groups, we need
to create a bootstrap token with a new option
"--groups=system:bootstrappers:kubeadm:default-node-token", to be able
to allow worker nodes to join the master node.

For that, introduce a new option TokenGroupsOption in the cluster
config, which is set only for k8s v1.8 or newer.