pkg/machinetool, pkg/nspawntool: pass opts to machinectl shell
When running machinectl shell with an env variable on, the variable is
not correctly passed, so that every command fails right away.
To make it work, we need to pass a new param opts that holds an
additional option to be passed. Then we can set opts to
`--setenv=KEY=VAL` which will work fine.