cmd/kube-spawn: check if image exists before removing it
Since kube-spawn stop & destroy always reported error messages when
the images do not exist, we need to check if the image already exists,
before removing the image.

For that, we need to introduce a new helper `machinetool.ImageExists()`
which is a simple wrapper of `machinectl show-image`. And change
several existing call helpers to make them also call
`machinectl.ImageExists()`.