cmd/kube-spawn: ignore resizing storage pool if the pool image does not exist
We should not resize the storage pool, if the pool image
`/var/lib/machines.raw` does not exist. In that case, it's just fine to
ignore resizing the image, because it means that systemd-machined did not
have to create an image to be loopback mounted.

Fixes https://github.com/kinvolk/kube-spawn/issues/170