bootstrap: search for `cni-noop` in `PATH` too
`kube-spawn` used `ForkExec` function from `syscall` package to run
`cni-noop` for setting up a bridge.  The `ForkExec` function is
basically a wrapper around `fork` and `execve` syscalls. And the
`execve` syscall does not check the PATH environment variable, which
in the end made `kube-spawn` to expect `cni-noop` to be in the current
working directory. Which usually won't be the case when `kube-spawn`
is installed somewhere in `PATH`, alongside with other `cni-noop` and
other binaries.

Just use `os/exec` package. It's simpler and there is a convenience
function that checks the PATH environment variable when searching for
the binary to be executed.