pkg/cnispawn: drop global variables
`goPath` global variables was unused, so just drop it.

`cniPath` global variable was set only in `Spawn` function. But this
package also provides a `NewCniNetns` function, which uses the
`cniPath` variable. The `NewCniNetns` function is used by `cni-noop`
which does not call the `Spawn` function at all, leaving the `cniPath`
variable empty.

Replace the variable with function that get the paths from the
`CNI_PATH` environment variable and use it in the `Spawn` function.