pkg/nspawntool: Search for cnispawn in `PATH`
The code was constructing the `Cmd` struct manually, without searching
for the binary in the `PATH` environment variable. The fix is to use
`Command` function which does that for us. This function also allows
us to stop repeating the name of the binary twice, so this allowed me
to simplify argument handling a bit.

There is also no need for setting the environment explicitly with
`os.Environ()` - it is the default behaviour.