pkg/utils: Add Command convenience function
The `Command` function from `os/exec` sets some internal error if the
desired binary is not found in `PATH` and this error is returned when
we actually try to execute the binary. We used this function to
simplify the way we execute some binaries, but the behaviour described
above is not backward-compatible.

This commit adds a similar function that creates a Cmd instance, but
tries to be backward-compatible - it prefers the binary in current
working directory over those in PATH.