cmd/kube-spawn: retry to remove image until it succeeds
Sometimes `kube-spawn stop` succeeded, but afterwards kube-spawn is
not able to remove `kubespawndefault*` images, because of errors like
`device or resource busy`. That makes the next round of starting
clusters fail because of the remaining images.

This happens simply because it takes some time until nspawn containers
can release every resource. So we need to retry to remove image until
it succeeds, up to 15 * 0.5 sec = 7.5 sec.

Fixes https://github.com/kinvolk/kube-spawn/issues/213