list: indicate when environments are invalid
Before kube-spawn supported multiple cluster environments,
/var/lib/kube-spawn used to contain directory such as "etc", "k8s" etc.
When updating to a newer kube-spawn, this is mixed with the different
environments. This makes the "kube-spawn list" command confused.

This patch makes it more obvious that "k8s" is not a real cluster name:
```
$ sudo -E kube-spawn list
ENV NAME            LAST MODIFIED
default             Nov 16 17:43:05
etc [INVALID]       Oct  9 16:39:36
extras [INVALID]    Sep 22 15:03:21
k8s [INVALID]       Oct  9 16:39:36
scripts [INVALID]   Sep 22 16:31:10

5 environment(s) found
```

To achieve this, the command list actually parses the kspawn.toml
configuration file and prints "[INVALID]" when the parsing fails. The
LoadConfig() function in 'pkg/config/utils.go' was changed to accept the
cluster name as parameter instead of using the viper flag. I think it's
better if code in 'pkg/' does not use viper in this way because
libraries should not have a clear API without hidden side-channels.
Besides, it will help when implementing a kube-spawn daemon with a gRPC
interface.