Fix dev workflow
Add --hyperkube-tag (or -t) on kube-spawn create to specify which
hyperkube Docker image to use. This is useful to compile different
versions of Kubernetes and test them in parallel. doc/dev-workflow.md
explains in more details how to use it.

kube-spawn will start a registry container and push the hyperkube image
there under the name 10.22.0.1:5000/hyperkube-amd64:$TAG. Then, kubeadm
will pick that image.

Use the unifiedControlPlaneImage option in the kubeadm configuration
file (>= 1.8) in addition to the old environment variable
$KUBE_HYPERKUBE_IMAGE (< 1.8). This means --dev cannot be used anymore
on 1.7 branches, but the 1.7 releases still work so it should not be a
problem.

The dev workflow is now documented in doc/dev-workflow.md

Fixes https://github.com/kinvolk/kube-spawn/issues/188