pkg/config: fix a bug with kube-spawn-dir
We need to define a toml attribute "dir" for kube-spawn-dir, to be able
to make use of a global option `-d' (or `--dir').
Otherwise the dir string is not stored in `kspawn.toml` at all, and
as a result, the custom directory name cannot be set.

Fixes https://github.com/kinvolk/kube-spawn/issues/238