pkg/utils: figure out kubeconfig path correctly with a cluster name
Now that kube-spawn is able to deal with multiple clusters, path to
kubeconfig, $KUBECONFIG, can also differ depending on the cluster name.
So `utils.GetValidKubeConfig()` should take a parameter for the cluster
name, to compose path to kubeconfig correctly. For example, if the
cluster's name is `mycluster`, KUBECONFIG will be
`/var/lib/kube-spawn/mycluster/kubeconfig`.

Also handle the case of a custom directory instead of
/var/lib/kube-spawn.

Fixes https://github.com/kinvolk/kube-spawn/issues/230