pkg/config: split out version checking function into IsVerConstraintOrDev
To able to avoid error-prone version checking logic, we need to
split it out into `IsVerConstraintOrDev()`. The function returns true
if it's dev cluster, or if the kubernetes Version is "latest"
(non-semver string), or the kubernetes version sastifies the given
constraint. That way we can simplify call sites as much as possible.

Fixes https://github.com/kinvolk/kube-spawn/issues/243