Remove cni-noop and `EnsureBridge`
`cni-noop` doesn't seem to be needed for a functioning kube-spawn.
There's no documentation as to why it was added in the first place.

Since containers are setup with `cnispawn`, they already should be in a
new network namespace at that point or have returned an error before
`EnsureBridge` is reached. Also, since `cni-noop` never calls
`netns.Set()`, it's not clear what the purpose of the "fallback" in
`EnsureBridge` actually is.

Remove `EnsureBridge` as well: from the code it's not clear why the
extra check was added and we should encounter an error earlier in case
setting up the bridge was not successful.