Replace cnispawn with hidden command `cni-spawn`
Cobra has a feature to hide sub commands from users, so `cni-spawn` can
be a subcommand and lets us avoid to build and deploy multiple binaries
(and with that get rid of the awkward setup instructions and
requirements in a later commit).

Also, get rid of the `background` option, which is not needed currently
(kube-spawn always uses the `background` way).