Remove requirement on `CNI_PATH` env var
Currently, users have to set a `CNI_PATH` environment variable pointing
to a CNI plugin dir in order to use kube-spawn. This is tedious as it
has to be done for every shell/environment and error prone (there's a
whole, albeit outdated, section in the troubleshooting document
dedicated to it).

We start with a simple change and ask the user to install the plugins
into a default location (or to provide the path via `--cni-plugin-dir`.)
This has the advantage that it's a one time task for most users and that
they don't have to prepare every session / environment.

The right thing to do would be to download the CNI plugins for the user,
cache them and use/copy them for every cluster as needed. For that,
I think refactoring the kube-spawn's caching should come first, hence
that for a later commit.