Another fix for overlay check
As a part of the check of the overlayfs support we are running stat on
the kube-spawn directory to determine the filesystem it is on. The
stat will fail if the directory does not exist. And that happens when
we have a clean state, that is - we remove the `kube-spawn` directory
(usually `/var/lib/kube-spawn`).