Fix reading flags from viper
`viper.BindPFlags` seems to be a function that does an eager reading
of values from `pflag.FlagSet`, so it needs to be called after the
parameters were actually parsed by `pflag` (or `cobra`), not when we
specify the flags we expect. The `PersistentPreRunE` hook seems to be
a fine place for this - the flags are already parsed and it is called
for all the commands.

There is no need to bind persistent flags separately - they are a part
of a result of the `cmd.Flags()` function invocation, and we pass the
result of `viper.BindPFlags`.