pkg/cluster: change API to v1alpha2 for kubeadm >= 1.11
Since k8s v1.11, API version of kubeadm has changed to v1alpha2, and
v1alpha1 is deprecated. So it's recommended to use v1alpha2 for
kubeadm >= 1.11, and v1alpha1 will be completely removed in the upcoming
release 1.12. We need to introduce a new option for the API version, to
conditionally deal with it.

Fixes https://github.com/kinvolk/kube-spawn/issues/280