Resolve errors where /var/lib/machines is not loopback mounted.
The code had an inherent assumption that /var/lib/machines
was loopback-mounted from /var/lib/machines.raw, and that it
could have exclusive access to this. Neither assumptions is
true in general. Additionally, it unconditionally would
umount/resize/re-mount, even if the volume was large enough.
If the volume were larger than needed, this had the effect
of trying to shrink it with qemu-img.

Add check that /var/lib/machines is large enough by calling
statfs. If it is, do no work.

If the volume is not large enough, attempt to call
machinectl set-limit. This is the documented method to use.

If this fails, then attempt to umount/resize/remount.