Add error handling to utility functions, remove assumptions about host path.
Most error checking was missing from pkg/bootstrap/node.go.
Added to all utility functions.

In addition, pkg/bootstrap/node.go made assumptions about
host-path for binaries (machinectl/qemu-img/mount etc). Remove this.
If the binaries are not on the PATH, fail. This is the reason
for Unix PATH, to indicate where to find executables, hard-coding
fallbacks can create a security issue or astonishment.

The code had a hard-coded assumption that the machine had
selinux available (e.g. it assumped selinux is enforcing/not-enforcing)
where there is a 3rd-option (not available). If selinux is not
available (e.g. Ubuntu-derived hosts), then return false
from isSELinuxEnforcing().