Put /var/lib/kubelet onto host as bind-mount rather than in container.
The /var/lib/kubelet directory is used to store the emptyDir
volumes of the various Pods. The worker container has < 1GB of
space free, insufficient. Change /var/lib/kublet to be
a bind-mount of the host.

Kubeadm reset removes the /var/lib/kubelet directory (which
is our bind-mount). However, kubeadm reset is not required
since the image starts up fresh, so remove it.