Enable ipv4/ipv6 dual-stack in Calico.
Calico is capable of enabling ipv4+ipv6 dual stack
usage, but only if it has an ipv6 range in ipam. Without the ipv6
range, even lo is not bound to a v6 address, preventing code
which explicitly wants to use v6 even locally (e.g. unit tests)
from operating.

Add in an explicit v6 range (from the the unique local address
range of fc00::/7). Do this by mirroring the config file from

 https://docs.projectcalico.org/v3.1/getting-started/kubernetes/installation
 /hosted/kubernetes-datastore/calico-networking/1.7/calico.yaml

into the system, with the change. The ipv4 section of the ipam
is configured to match Calico's felix expectation.