pkg/cluster: set COLUMNS to a bigger value to get systemd machine names
Since systemd v239, `machinectl {list,list-images}` started to shorten
each machine name with ellipsis("…"), when the terminal is narrow and
unicode is being used.

So before v239, `machinectl list` showed like this:

```
kube-spawn-default-master-4nr5ar container systemd-nspawn - - 10.22.0.206...
```

But since v239, it's like that:

```
kube-spawn-default-master-4n… container systemd-nspawn - - 10.22.0.206…
```

As a result, each machine name returned by `List()` and `ListImages()` does
not match with an expected name. Both `kube-spawn start` and
`kube-spawn stop` do not work, when unicode is being used.

There does not seem to be a way to avoid the shorted name by using an
option. Adding `--full` does not solve anything.

To fix that, we need to set an environment variable `COLUMNS` to a
fairly big value, 200.