pkg/cluster: apply CNI plugins outside of nspawn containers
Starting from k8s v1.13, the insecure port 8080 of kube-apiserver
is not available any more, as it has been deprecated. So we cannot
run `kubectl apply` inside of nspawn containers, unless some additional
RBAC config would be done for allowing localhost to access to the secure
port.

Let's take a different approach. Apply CNI configs outside of the
containers. Simply get path to kubectl & kubeconfig, and run on the
host. To do that, we need to move call site of `applyNetworkPlugins()`
to later, because immediately after `kubeadmInit()`, the API server on
the master node is not ready yet. Instead we need to do it later, after
`kubeadmJoin()`, when the API server is available from the host.
Otherwise `applyNetworkPlugin()` definitely fails.